refactor: split model/labelkey file
