+++ pkg/model/model.go
@@ -15,7 +15,6 @@ import (
 	"github.com/charmbracelet/bubbles/textinput"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
-	"github.com/lithammer/fuzzysearch/fuzzy"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/fields"
@@ -23,57 +22,6 @@ import (
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/tools/clientcmd"
 )
-
-/* LabeKey */
-type LabelKey struct {
-	Name  string
-	Style lipgloss.Style
-}
-
-func NewLabelKey() *LabelKey {
-	return &LabelKey{}
-}
-
-func (k *LabelKey) WithName(name string) *LabelKey {
-	k.Name = name
-	color := hashToColorCode(hash(name))
-	style := lipgloss.NewStyle().Foreground(lipgloss.Color(color))
-	k.Style = style
-
-	return k
-}
-
-func (k *LabelKey) Render() string {
-	value := ""
-	style := k.Style.Copy()
-
-	if len(k.Name) > searchModelWidth {
-		value = k.Name[:searchModelWidth-3] + "..."
-	} else {
-		value = k.Name
-	}
-
-	return style.Render(value)
-}
-
-/* LabeKeys */
-func LabelKeyNames(keys []LabelKey) []string {
-	var names []string
-	for _, key := range keys {
-		names = append(names, key.Name)
-	}
-	return names
-}
-
-func FuzzyFindLabelKeys(input string, keys []LabelKey) []LabelKey {
-	var results []LabelKey
-	for _, key := range keys {
-		if fuzzy.Match(input, key.Name) {
-			results = append(results, key)
-		}
-	}
-	return results
-}
 
 /* LabelValue */
 type LabelValue struct {
